41: Refactor annotation helpers
